Update gutFreqWaveSpeedFinder.m
Check that filter creation doesn't give an error based on dataset size; expand if needed. Note that the size of LPFilt.Coefficients gives the order of the filter, and the dataset must be 3X this.
Offset dTimeMin to be 1/fps (at least)
Improve filtering, etc., to deal with variation in frame rate. Alter filter sizes to depend on frames-per-second:
% Make 'PassbandFrequency',  and 'StopbandFrequency' depend on fps. Note that values of 0.15 and 0.65 seem fine for 5 fps, so scale by this; max 1 (normalized frequency)
LPFilt=designfilt('lowpassfir', 'PassbandFrequency', min(0.15*5/fps, 1.0), ...
        'StopbandFrequency', min(0.65*5/fps, 1.0), 'PassbandRipple', 1, ...
        'StopbandAttenuation', 60);
Minor output formatting changes